feat(alert-echoes-timeline): rename component and integrate new timeline service
