+++ frontend/src/app/data-management/alert-management/shared/components/alert-echoes-timeline/alert-echoes-timeline.component.html
@@ -2,6 +2,7 @@
   <!-- Main ECharts timeline -->
   <div echarts
        [options]="chartOption"
+       (chartInit)="onChartInit($event)"
        class="timeline-chart"
        style="height: 400px; width: 100%; margin-bottom: 10px;"
        (chartClick)="onChartClick($event)">
