+++ frontend/src/app/data-management/alert-management/shared/components/alert-echoes-timeline/alert-echoes-timeline.component.scss
@@ -21,7 +21,7 @@
 
   .timeline-chart {
     width: 100%;
-    height: 400px;
+    height: 500px;
     margin-bottom: 15px;
     border: 1px solid #ddd;
     border-radius: 4px;
